requirements.txt
# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.23
langchain-text-splitters>=0.0.1
langgraph>=0.0.20
openai>=1.12.0
tiktoken>=0.5.2

# Vector Databases
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Email
sendgrid>=6.11.0
aiosmtplib>=3.0.1
email-validator>=2.1.0

# Flask API
flask>=3.0.0
flask-cors>=4.0.0
flasgger>=0.9.7.1
gunicorn>=21.2.0
waitress>=2.1.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
PyYAML>=6.0.1
python-multipart>=0.0.6
rich>=13.7.0
typer>=0.9.0

# Optional: PDF support
pypdf>=4.0.0

# Optional: Advanced document loaders
unstructured>=0.10.0
